+++ ui_automation_tests/pages/application_page.py
@@ -23,7 +23,6 @@ class ApplicationPage(BasePage):
     PROGRESS_APP_BTN = "change-status"  # ID
     ACTIVITY_CASE_NOTE_SUBJECT = ".app-activity__list .govuk-body"
     ACTIVITY_DATES = ".app-activity__item .govuk-hint"
-    ACTIVITY_USER = ".user"
     IS_VISIBLE_TO_EXPORTER_CHECKBOX_ID = "is_visible_to_exporter"
     REVIEW_GOODS = "button-review-goods"
     EDIT_CASE_FLAGS = "application-edit-case-flags"  # ID
@@ -80,12 +79,6 @@ class ApplicationPage(BasePage):
     def get_text_of_case_note_date_time(self, no):
         return self.driver.find_elements_by_css_selector(self.CASE_NOTE_DATE_TIME)[no].text
 
-    def get_disabled_attribute_of_post_note(self):
-        return self.driver.find_element_by_id(self.BUTTON_POST_NOTE_ID).get_attribute("disabled")
-
-    def get_class_name_of_post_note(self):
-        return self.driver.find_element_by_id(self.BUTTON_POST_NOTE_ID).get_attribute("class")
-
     def click_progress_application(self):
         self.click_drop_down()
         scroll_to_element_by_id(self.driver, self.PROGRESS_APP_BTN)
@@ -120,9 +113,6 @@ class ApplicationPage(BasePage):
 
     def get_text_of_activity_dates(self, no):
         return self.driver.find_elements_by_css_selector(self.ACTIVITY_DATES)[no].text
-
-    def get_text_of_activity_users(self, no):
-        return self.driver.find_elements_by_css_selector(self.ACTIVITY_USER)[no].text
 
     def click_review_goods(self):
         self.driver.find_element_by_id(self.REVIEW_GOODS).click()
