+++ ui_automation_tests/pages/shared.py
@@ -11,12 +11,11 @@ class Shared(BasePage):
     LITE_TABLE_ROW = ".govuk-table__body .govuk-table__row"  # CSS
     LITE_TABLE_CELL = ".govuk-table__body .govuk-table__cell"  # CSS
     GOVUK_CAPTION = ".govuk-caption-l"  # CSS
-    SELECTED_TAB = ".lite-tabs__tab--selected"  # CSS
     BODY = "body"  # CSS
     LINKS_IN_TABLE = ".govuk-table td a"
     ROWS_IN_CASES_TABLE = ".govuk-table__body .govuk-table__row"  # CSS
-    H2 = "h2"  # CSS
     LINKS_IN_LITE_TABLE = ".govuk-table__cell a"  # CSS
+    INFO_BAR = ".lite-info-bar"  # CSS
 
     def click_submit(self):
         self.driver.find_element_by_css_selector(self.SUBMIT_BUTTON).click()
@@ -27,9 +26,6 @@ class Shared(BasePage):
     def get_text_of_body(self):
         return self.driver.find_element_by_css_selector(self.BODY).text
 
-    def get_text_of_selected_tab(self):
-        return self.driver.find_element_by_css_selector(self.SELECTED_TAB).text
-
     def get_text_of_caption(self):
         return self.driver.find_element_by_css_selector(self.GOVUK_CAPTION).text
 
@@ -38,9 +34,6 @@ class Shared(BasePage):
 
     def get_text_of_lite_table_body(self):
         return self.driver.find_element_by_css_selector(self.LITE_TABLE_BODY).text
-
-    def get_text_of_h2(self):
-        return self.driver.find_element_by_css_selector(self.H2).text
 
     def get_links_in_cells_in_table(self):
         return self.driver.find_elements_by_css_selector(self.LINKS_IN_TABLE)
@@ -68,3 +61,6 @@ class Shared(BasePage):
         edit_buttons = self.driver.find_elements_by_css_selector(self.LITE_TABLE_ROW)
         row_index = str(len(edit_buttons))
         scroll_to_element_by_id(self.driver, "row-" + row_index)
+
+    def get_text_of_info_bar(self):
+        return self.driver.find_element_by_css_selector(self.INFO_BAR).text
