+++ ui_automation_tests/step_defs/test_give_advice.py
@@ -72,9 +72,7 @@ def import_text_advice(driver, option, context):
 
 @then("I see my advice has been posted successfully")
 def posted_successfully_advice(driver):
-    pass
-    # Commenting the below out due to bug.
-    # assert Shared(driver).get_text_of_info_bar() == "Your advice has been posted successfully"
+    assert Shared(driver).get_text_of_info_bar() == "Your advice has been posted successfully"
 
 
 @then("I see added advice in the same amount of places")
